requirements.txt
numpy
statsmodels
seaborn
scikit-learn
pytorch
tqdm